requirements.txt
Flask
Flask-Cors
google-generativeai
python-dotenv
gunicorn
